Fix wrong START offset
Problem 1: START is set before looking up zip64 extra field. This
can lead to unproper LHF offset in case where the first entry is
located further than 0xFFFFFFFF in the archive.

Problem 2: If the order of entries in the CD does not match the
order of the LFH (e.g: The first CD entry is not the first LFH in
the archive), file scanning discard previous entries and mark them
as prefix.

Solution: Leverage @CD to generate a better START offset.